src/main/java/study/datajpa/repository/MemberJpaRespository.java
package study.datajpa.repository;

import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import org.springframework.stereotype.Repository;
import study.datajpa.entity.Member;

/**
 * packageName    : study.datajpa.repository
 * fileName       : MemberJpaRespository
 * author         : kmy
 * date           : 10/27/23
 * description    :
 * ===========================================================
 * DATE              AUTHOR             NOTE
 * -----------------------------------------------------------
 * 10/27/23        kmy       최초 생성
 */
@Repository
public class MemberJpaRespository {
    @PersistenceContext
    private EntityManager em;

    public Member save(Member member){
        em.persist(member);
        return member;
    }

    public Member find(Long id){
        return em.find(Member.class, id);
    }
}
